Minor updates
Minor update in the beginning. Not sure if the same text is used as well with other client documentation. @Liza; Can you update in the other sections?

Update to Admin Session: I think we do not need to mention here Server 2003 anymore. I couldn't find TechNet documentation on the console session. I wonder if you have somewhere in the archive that can be reused.