Update monthly-delta-update-isv-support-without-WSUS.md
Fixed references that said "Windows 10 1704" as it's actually 1703.  Changed to use the primary IT pro labels, e.g. "Windows 10 1703" with the marketing name in parenthesis, e.g. "Windows 10 1703 (Creators Update)".  We want to focus IT pro documentation on the YYMM names.

Also note that there is a typo in the graphic used on the page, misspelling the word "Quality."  I don't have any way of fixing that :-(