Mention new database creation
When upgrading AD FS FBL a new Adfs Configuration database is created. This is worth mentionning this can have some effect on the infratructure setup. In our case the SQL admins didn't realise this and the ADFS admins didn't get this from the docs. In the end the configuration database was not configured for high availability which caused some unplanned downtime.