Remove duplicated section (fixes #319)
As noted in #319, the "Are third party proxies supported with AD FS?" FAQ entry exists twice (under both Deployment and Design). The duplicate was added in 6892d0ccd6ffcbef2f3de30f4a628b719d6cc7d1. Removing that one as it seems to be better suited with the other Design questions.