Update What-s-new-in-Hyper-V-on-Windows.md
Added warning not to try to use the tech for backing up clusters on single VM's - this was due to a request from Ben Armstrong, based on a  conversation he's had with some of his customers. Unfortunately, I don't know of any place else to put this - if anyone has a better location, I'd be happy to move it there.