Merged PR 9236: removed article from repo
Removed article from repo. Article was previously redirecting, but because it was still in the repo, customers were able to find the information in the public GitHub repo, leading to confusion about whether the feature was supported (it's not). Removed the article completely + added global redirect to the redirection.json file.