Fixed code markdown
Some of the PowerShell code in the article was showing incorrect backsticks (`) in the final render in docs.microsoft.com. Fixed the markdown for these commands and added the language name to allow color coding.
Removed some unnecessary whitespaces that were causing that the numbered list was not rendered properly.
Added backsticks to text that should treated as code and not automatically translated (e.g. the reference to the file Style.css).