Realigned this page to an earlier incarnation
For some weird reason someone had a field day on this table and added weird spaces, changed the implication of some explanations (to incorrect ones) and made it generally hard to know what was going on. It should be reviewed by someone who knows exactly how the regex engine of NPS works. This is the earlier incarnation of the page that I used to correct some of the failings
https://docs.microsoft.com/en-us/previous-versions/windows/it-pro/windows-server-2008-R2-and-2008/cc755272(v=ws.10)