Trying again to fix the broken requirements
We've been through this on 2016, but I'm hoping we can correct this for 2019.  The official requirements mean you officially can't run Windows Server 2019 in a VM (because memory type isn't known), and you can't run in Azure (because PXE isn't available.)  Those are both patently false statements.  In the past, for 2016, the statement was made that "we'll make special requirements for Azure," which was never done and was a very broken idea to begin with.  Can we please get this accurate this time?  (There's other minor issues here I'm not even worrying about right now - these are bigger ones.)