Admin Center - FSRM Quotas in Storage not Files
To edit or update a quota, one needs to open the Storage node - not the Files node.

In addition, it is misleading to suggest that all FSRM options are available. Setting quota management, denying disk space to users, and limiting disk usage is all that is available. Modifying FSRM e-mail templates, SMTP settings, or any other additional FSRM feature has not been added yet.