Remove brackets from parameters table
Per Loc bug https://mseng.visualstudio.com/TechnicalContent/_queries/edit/1537584/?triage=true, there's an extra bracket in the first row, [/remote:], but my understanding of the syntax is that either all of the parameters should have brackets (because all are optional), or none should. So I removed brackets from the few entries that had them, but please check that this is correct.