Fix inverted logic in example section
In the "Create a zone level zone transfer policy," the description does not appear to match the command. The description indicates that the command will IGNORE any zone transfers from the given server interface IP. However, since the command used "ne," the inverse was true - the DNS server would ignore all zone transfer requests from all interface IPs except the specified one.

To correct this, I've inverted ne to eq. In my lab, this appears to be the correct command.