Virtualization/Hyper-V: PowerShell parameter SMB
Description:

As concluded in issue ticket #2623 (The PowerShell command says
"SMBTransport" but that generates an error), the PS command argument
parameter "SMBTransport" does not work because it is incorrect.

Proposed changes:

- replace "SMBTransport" with only "SMB"
- remove redundant white space at the end of lines

Thanks to VernAnderson for reporting the issue and to Tajmaha
for confirming that the simplified parameter works as intended.

issue ticket closure or reference:

Closes #2623