Upgrading to AD FS: typos and whitespace
Description:
As pointed out in the issue ticket #2707 (Mistype ?), there is a typo in
the sentence "[...] note that this requires one of the new Windows
Server 2019 federation servers tp be promoted to the role of primary
node in the farm." -- where "tp" should have been "to".

Thanks to TheOtherBillGates and NomadRex for reporting this typo.

While browsing this document source to correct this typo, I noticed
a couple of other typos as well, and later on noticed a lot of
whitespace inconsistencies. I have replaced some occurrences of
"commandlet" with 'cmdlet' (the latter being the proper PS term).

Proposed changes:

- correct a few typos (tp, rasing)
- replace "commandlet" with 'cmdlet'
- remove trailing whitespace (blanks at the end of lines)
- add missing MarkDown quote marker compatibility space in 2 Notes
- remove 3 redundant blank lines
- keep PowerShell code blocks preceded and succeeded by a blank line
- move 2 cmdlets to a proper code block with PS syntax highlighting

issue ticket closure or reference:

Closes #2707