IP Address Management/broken link: Server Cmdlets
Description:

Based on community feedback in issue ticket #3477 and #608, the link to
IP Address Management (IPAM) Server Cmdlets in Windows PowerShell
is broken again, if it was ever fixed back in April 2018.

Thanks to vicslive and Verox- for reporting this issue, even though
this did not get noticed before today in the new issue ticket (#3477).

Changes proposed:
- replace the broken TechNet link with a working docs.microsoft.com one
- clear up whitespace for codestyle adherence

issue ticket closure or reference:

Closes #3477