Remote Access/IP Helper service: PS cmdlet typo
Description:

As described in issue ticket #3568 (It appears there is a spelling
mistake on the page.), the service name IPHlpSvc is misspelled in the
PowerShell cmdlet to use for checking if the service is running.

Thank you to termanix for finding and reporting this typo.

Proposed changes:

- add the missing letter v to "iphlpsc" in the cmdlet
- add the MD syntax highlight keyword PowerShell to the MD code block
- remove the empty whitespace lines at the end of the document

issue ticket closure or reference:

Closes #3568