Deploy DHCP/PowerShell: convert technet links
Description:

As mentioned in issue ticket #3133 (Many links still pointing to
Technet), most of these links can now be found in the
docs.microsoft.com directories instead, most of them in the
/powershell/ subdirectory.

Thanks to Marc Laflamme (MarcLafLigado) for pointing out this rather
obvious need for the links to be updated, ticket opened half a year
ago, on July 22, 2019. A couple of links were even broken already
(local MS error 404), but with suggested search results
pointing to the new pages. Thanks to ingd for reminding us.

Changes proposed:
- convert technet.microsoft.com links to docs.microsoft.com links
- add MarkDown compatibility spacing for MD Note blob indents

issue ticket closure or reference:

Closes #3133