ADFS/Deployment: resurrect 2 internal links
Description:

As described in issue ticket #2863 (Broken Links), the first 2 links in
this article use an outdated format of
assetId:///e340cf8f-acf3-4cba-8135-a9353b85e714
which was supposed to point to the article itself and 2 sections inside
this article by means of anchor links (#BKMK_1 & #BKMK_2).

Thanks to Sean Harper (harperse) for creating the ticket to point out
this issue in the first place. Thanks also to fate0521 for reminding us.

Changes proposed:
- Replace the outdated assetId:/// links with filename + anchor
  (as used in the anchor link BKMK_3 below).
- Remove whitespace (blanks) at the end of all lines.

issue ticket closure or reference:

Closes #2863