Update delimit-volume-allocation.md
The minimum number of servers for a delimited volume is 4, not 3.  3 can only survive one copy being down due to quorum requirements of the space (which is separate from pool quorum).  To allow 2 nodes being down, it requires 4 servers.  I had to remove a lot of the graphics, there are no replacements and showed 3 servers instead of 4. For the same reason I had to remove some of the table/examples.