Windows Commands/Secedit: add missing slash
As pointed out in issue ticket #3690 (Missing backslash), the optional
command parameter /log is missing its slash in the Syntax line. The
slash is used in the example section below, so it must have been
forgotten in this Syntax section.

Thanks to Francisco Celedon (fceledon) for pointing out the typo.

Proposed change:
- add missing slash for the /log option

issue ticket closure or reference:

Closes #3690