Remote Desktop/client access: sitting not required
Description:

As pointed out in issue ticket #3706 (Need not be sitting), the
computer in question does not require the user to be sitting when
operating it, as long as the user is physically present at the device.

Thanks to kdheerendra for pointing out the unnecessary possibility for
misinterpretation in this sentence. There is no requirement to be
physically sitting, what so ever. :-)

Proposed change:
- Replace the phrase "sitting in front of" with "using".

issue ticket closure or reference:

Closes #3706