Windows Commands/bitsadmin getstate: restore table
Description:

As pointed out in the issue ticket #3703
(**Broken table and Missing State in Bitsadmin\getstate**),
the Remarks table was broken due to a missing header row and could
therefore not render properly, neither in the MarkDown source nor HTML.

This PR should resolve the table issue as well as clean up the page
source for future editing. Thanks to WiVi71 for reporting this issue.

Changes proposed:
- restore table structure by adding the missing header row
- clean up table formatting by adding spacing to make it more justified
- adding a couple of blank lines for editing clarity
- removing a redundant blank line after the page title

issue ticket closure or reference:

Closes #3703