PowerShell/Nano Server: 404 MSDN Link replacement
Description:

As pointed out in the issue ticket #3740 (**Broken Link**), the link in
the last line of this document points to a non-existing MSDN page and
therefore causes a 404 error (with suggested alternatives).

Thanks to estoces (Eric Stoces) for reporting this broken link.

Changes proposed:
- replace the MSDN link with a working docs.microsoft.com link
- remove all trailing (end-of-line) spaces

issue ticket closure or reference:

Closes #3740