AD FS OBO: secure WebAPIOBOController with Authorize attribute
Description:

As discussed in issue ticket #3061
(Shouldn't WebAPIOBOController be secured with an Authorize attribute?),
the WebAPIOBOController controller should be secured with an Authorize
attribute to validate the OAuth token when the ToDoListService calls
the WebAPIOBO API, otherwise the AD FS configuration for
WebAPIOBOController will never really be tested
when running this locally.

Thanks to Toby Artisan (tobyartisan) for making us aware of this issue.

Changes proposed:
- add the [Authorize] attribute within the AD FS OBO Controller code
- add user: {User.Identity.Name} to the output string
- add MarkDown syntax highlighting to the code snippet
- improve the grammar/readability of preceding instructions
- change the MD code block fence characters to back ticks

issue ticket closure or reference:

Closes #3061