#3466 AD DS/Virtual DC: replace old TechNet links
Description:

This PR concludes the job of cleaning up the unresolved and closed pull
requests #3463, #3464, #3465, #3466, and #3467.

This page was one of the few still in need of updating.

Thanks to @mikeamoss for the original work of finding the correct links.

Changes proposed:
- Replace 4 old technet links with docs.microsoft.com links
- Remove redundant whitespace (blank lines) at end-of-file
- Remove redundant whitespace (blank space) at end-of-line

To see the effective changes without whitespace changes, use the "Hide
whitespace changes" feature in the GitHub source file view page.

Ticket closure or reference:

Ref. #3466 (the PR is already closed)