Deploy Always On VPN: Restore Larger image Link
Description:

As pointed out in the issue tickets #3178 (image size very small)
and #3474 (Image in article is unreadable), the workflow process image
is too small for the text content to be read (even opening the image in
a new browser tab to enlarge it, the text is pixelated and blurry, so
one can barely guess what the text says anyway).

2 years ago, in 2018, there used to be a working HTML link from that
very image in this page to the larger version of the same image
(always-on-vpn-deployment-workflow.png), but some time during the last
2 years, this link has disappeared. This PR is an attempt to restore it.

Thanks to LeonB87 (Léon) and tomdan-MSFT for posting the issue tickets
to make us aware of this image readability issue.

Changes proposed:
- Restore HTML link via MarkDown, point the Small image (-sm) to Large
- Add MarkDown compatibility spacing to the [!IMPORTANT] blob markers

Ticket closure or reference:
Closes #3178
Closes #3474