In-Box VSS Writers: purge links to removed content
Description:

As pointed out in issue ticket #3947 (Broken Links), the links to "New
In-Box VSS Writers for Windows Server 2008 and Windows Vista SP1" and
"New In-Box VSS Writers for Windows Server 2008 R2 and Windows 7" are
now pointing to a page saying "Content Removed (This content has been
removed.)" The links are, respectively,

- https://go.microsoft.com/fwlink/?linkid=180896
- https://go.microsoft.com/fwlink/?linkid=180897

Thanks to SauerKraut51 for opening the ticket to point out this issue.

In addition, the remaining link "In-Box VSS Writers"
(https://go.microsoft.com/fwlink/?LinkId=180895) is now permanently
redirected from MSDN to the docs.microsoft.com site.

Changes proposed:
- Remove the 2 links, along with the title text
- Modify the preceding text to indicate that only 1 link will follow
- Replace the remaining link with the direct link to the MS Docs page
- Also indicate that this is the case by modifying the preceding text
- Restore the H3 heading MarkDown format title code before the title
  "Supported Operating System Versions" in row 517 (was 522)
- Add proper NewLine to the last line of the page

Ticket reference or closure:

Closes #3947