The Role of Claim Rules: reference correction
Description:

As pointed out in issue ticket #3992 (Edit recommended), there is a
copy-paste mistake in the second to last rule set description. The last
Description sentence for the Delegation Authorization Rule Set reads:
"Unless you specify an issuance authorization rule, no users can act as
delegates by default." This sentence is identical in its first part
(before the comma) to the start of the last sentence in the previous
section, Issuance Authorization Rule Set.

The obvious solution is to refer to the name of the rule set by
replacing the duplicated words "an issuance" with "a delegation".

Thanks to PishPoshery (David Sitner) for pointing out this issue.

Changes proposed:
- Replace the duplicated words "an issuance" with "a delegation".
- Whitespace changes: remove all redundant end-of-line blanks.

Note: Remember to use the "Hide whitespace changes" feature to see the
effective changes easier in this Pull Request.

Ticket closure or reference:

Closes #3992