Web Application Proxy/Publishing: outdated links
Description:

As reported in issue ticket #4000 (Broken link), the link text
"Configure a site to use Integrated Windows authentication" does not
work as a link to any other page or section on this page, because it
uses the outdated format assetId:/// which is no longer supported.

The complete link looks like this:
assetId:///b0788958-627f-450f-877c-209b1bd0db52

There are also 4 outdated technet links which I suggest replacing with
the current redirected docs.microsoft.com links, even if they are
no longer regularly maintained (and flagged as such).

Thanks to joshtransient for reporting the broken link so I could
discover and amend some of the other issues on this page.

Changes proposed:
- Replace the assetId link with a working MS Docs page link
- Replace 4 technet links with their permanent redirect links
- Add a missing comma (for readability) in one of the sentences
- Add MarkDown Syntax Highlighting for the PowerShell cmdlets
- Whitespace changes:
    - "Trim Trailing Space" (remove all blanks at end-of-line)
    - Add MarkDown indent compatibility spacing for "Applies to:"
    - Normalize bullet point list marker spacing to only 1 space

Remember to use "Hide whitespace changes" or Rich diff view to see the
effective changes and skip (most of) the whitespace changes.

Ticket closure or reference:

Closes #4000