In the section Configuration Information
We have stated that the command to use to view the node selected as User Activity master is Get-AdfsFarmInformation.FarmRoles and if we run the command as it is it will error stating "Get-AdfsFarmInformation.FarmRoles : The term 'Get-AdfsFarmInformation.FarmRoles' is not recognized as the name of a cmdlet, function, script file, or operable program. Check the spelling of the name, or if a path was included, verify that the path is correct and try again".  The revised command to be used is (Get-AdfsFarmInformation).FarmRoles