Admin/FTP: turn hidden 'put' parameters visible
Description:

Based on the feedback comment in ticket #4086 
(Excellent ftp reference), I discovered that this
page does not show the parameters in their table.
The angle brackets are the cause of this issue,
`<` and `>`, also known as the "Less Than" and 
"Greater Than" symbols, often indicate HTML tags 
in web page encoding. The table entries need to be 
escaped or encapsulated to avoid being interpreted 
as hidden code. Based on the syntax view, I will 
prefer to use MarkDown back tick encapsulation.

Thanks to pernak01 (Ken Perna) for the opportunity.

Changes proposed:
- Add MD back ticks around the bracketed table entries
- Add metadata details "FTP - put" for this command
- Whitespace changes:
    - Remove end-of-line whitespace (trim trailing space)
    - Add MD indent marker comp. spacing in "Applies To:"
    - Add alignment spacing and hyphens to the MD table.

Ticket closure or reference:

Closes #4086