Windows commands/bitsadmin: multi-page corrections
Description:

As reported in issue ticket #4137 (Some mistakes in bitsadmin /*),
there are a few issues in these 9 pages.
- Broken table (2 pages)
- Double commands (2 pages)
- Wrong examples (2 pages)
- Unexpected space (3 pages)

Thanks to WiVi71 for reporting these issues.

Changes proposed:
- Add missing MD table header (2 pages)
- Remove duplicate command (2 pages)
- Correct example parameter (2 pages)
- Remove unexpected space (3 pages)

- Formatting readability improvements:
    - Remove redundant blank lines below title header
    - Align MD table grid characters (readability/within reason)
    - Add recommended blank line before example MD code fence
    - Add missing NewLine before EOF (end-of-file)

Additional note:

Regarding the comment "And also value 3 is missing":

No, the 3 is a bitwise combination of 1 + 2, meaning
the combination of parameter flags 1 + 2 = 3.
Logically speaking, this implies the following list of options:

1 : Generate an event when all files in the job have been transferred.
2 : Generate an event when an error occurs.
3 : Generate an event for both instances above.
4 : Disable notifications.

If you want me to add that bitwise combination value & description to
the table for bitsadmin /SetNotifyFlags , please comment in the PR.
(Otherwise, I will refer to this as an obvious conclusion for coders.)

Ticket closure or reference:

Closes #4137