Job Aids for Windows Server 2003 Deployment Kit
Description:

As noted in PR #4167 (Selecting the Forest Root Domain: updated links),
I have been informed as follows: "We typically are trying to replace
the fwlinks with direct URLs as we go through files. If you happen to
see them and feel like it, you can just replace the whole fwlink with
the direct URL."

This PR aims to correct all the specific links to the download page for
"Job Aids for Windows Server 2003 Deployment Kit".

Thanks to eross-msft (Elizabeth Ross) for the valuable feedback.

I have noticed that many other links are either outdated or broken, but
I have not yet had time to extend my research to those other links yet.
If you know the correct new link address for any of the outdated links
in any file in this Pull Request, feel free to tell me about it,
so I may add that/those link correction(s) as well.

Changes proposed:
- Replace https://go.microsoft.com/fwlink/?LinkID=102558 with
  https://www.microsoft.com/download/details.aspx?id=9608 anywhere
- Change the page name reference text to link text for existing links,
  instead of the link being used both as link text and the link itself.
- Remove any redundant end-of-line (EOL) whitespace
- Add MarkDown indent marker compatibility spacing (1 space) anywhere
- Standardize bullet point list internal indent spacing from 3 to 1

Ticket closure and reference:

Closes #3592, #3760, #3964
Ref.: #4147, #4167