Diagnostics Hub Standard Collector service text
Description:

As reported in issue ticket #4164
(Microsoft (R) Diagnostics Hub Standard Collector),
the service description text for this particular service on this page is
copy-pasted from the preceding service "Local Session manager".

Thanks to rlappsupport for reporting this documentation issue.

Changes proposed:
- Replace the current service description text with the correct one:
  "Diagnostics Hub Standard Collector Service. When running, this
   service collects real time ETW events and processes them."
-  Correct its name to "diagnosticshub.standardcollector.service"
- Correct its Startup type to "Manual" according to factual standard
- Rename row names in accordance with actual services.msc names
  (Service name, Description, Startup type)
- Swap order for the "Description" and the "Service name" rows
- Whitespace & Layout changes:
    - Remove all redundant end-of-line (EOL) whitespace (blanks)
    - Remove redundant blank lines at end-of-file (EOF)
    - Reduce blank line spacing to max 2 blank lines (from 3 or 4)
    - Standardize table layout spacing and borders of first column
    - Standardize cell layout for the blank header row (readability)

I realize that this PR contains a lot of changes and that it may take
more time than usual for MS Docs team members to review it.
My motivation for making so many changes is for this documentation
to be as useful and valuable as possible, for both myself and others,
using Microsoft Windows Server 2016 in current production scenarios.

Ticket closure or reference:

Closes #4164