Windows Server Insider Preview download page link
Description:

As reported in issue ticket #4186
(Broken link: Windows Server Insider Preview download page),
the current download link to the insider preview edition of
Windows Server returns the following error message: "We are sorry, the
page you requested cannot be found. The URL may be misspelled or the
page you're looking for is no longer available."

That error is caused by leaving out the locale /en-us/ part of the link.
This locale part is a unique requirement for this link to be working.
This PR corrects that issue in 2 files, along with other adjustments.

Thanks to danblakely for reporting this issue.

Changes proposed:
- Add the missing locale /en-us/ to the Windows Server Insider Preview
  download link
- Update 3 links from the MSDN format to current docs.microsoft.com
- Modify the [Perfmon API] link to [Using Performance Counters] due to
  current changes in the MS Docs library structure
- Whitespace changes:
    - Add missing MarkDown indent marker compatibility spacing (19 rows)
      (50% of those adjustments are MD table whitespace additions)
    - Remove redundant end-of-line blanks (1)

Ticket closure or reference:

Closes #4186