ADFS/Custom Id Tokens: New App Link & formatting
Description:

During an attempt to find out if this page contains answers to the
issue tickets "Not Work" #4178 and "E-mail and name claims not working"
#4235, I found several text & formatting problems worth resolving, as
well as discovering that the OpenID Web App on GitHub has been archived
and has therefore been replaced by a new version, also on GitHub.
Therefore, the link needs to be updated to point to the new repository.

I also found it necessary to perform some whitespace cleanup throughout.

Changes proposed:
- Update the OpenID Web App link to point to version 2 of the app
- Normalize the indents for the MarkDown code block fences
- Add MD syntax highlighting names for the code blocks: C# (cs) and XML
- Normalize whitespace use within bullet point lists and numbered lists
- Grammar particles added (the) as sentence structure improvements
- Add MD back ticks around code references for monospaced typeface
- Add grammatical commas, for readability and sentence structure
- Add & remove blank lines, based on general MD layout principles

I don't know the status of this page, if there are any plans to retire
the document or replace it with a new one. I simply thought that it
would be practical if it would be possible to make the content useful
by correcting the most obvious MarkDown formatting issues here.

Ticket closure or reference:

Ref.: #4178 #4235