Windows Server upgrade Overview: grammar details
Description:

As suggested in issue ticket #4265 (Suggested change to Title), the
current title text "Overview about Windows Server upgrades" is not
entirely written as in normal, everyday English usage. "Overview about"
is not a normal English phrase, compared to "Overview of".

There are also a few similar grammar details to be corrected in this
document to make it easier to read, understand and translate.

Changes proposed:
- "Overview about" -> "Overview of", in the title & metadata
- "think about" -> "think through" in the metadata section
- Remove a redundant "that's" from "on any Windows Server configured to"
- Remove an incorrect duplicate "be" from "can be typically be upgraded"
- Whitespace adjustments:
    - remove end-of-line (EOL) blank spaces
    - add MarkDown indent marker compatibility spacing
      in the [Important] note

Ticket closure or reference:

Closes #4265