Security/TLS: Link updates, whitespace & typo fix
Description:

As reported in issue ticket #4289 (Wrong GPO Path), there is an
unwanted s at the end of the GPO directory name "Network".

Thanks to matziq for reporting this typo.

There is also a fair amount of old MSDN & Technet links, with a
permanent redirect to docs.microsoft.com, in need of updating.

Changes proposed:
- Typo correction, "Networks" -> Network
- All  MSDN & Technet URLs updated to current targets
- Remove end-of-line blanks a.k.a. "Trim Trailing Space"
- HTTP to HTTPS (http://www.iana.org -> https://www.iana.org)
- Reduce redundant spacing in numbered list, from 4 down to 1
- Add MarkDown indent marker compatibility spacing in "Applies to:"
- Remove 1 blank line at EOF

Ticket closure or reference:

Closes #4289