Windows Commands: MD corrections for NFS table
Description:

As reported in issue ticket #4316 (Table not clearly displayed), the
MarkDown table is unreadable because the MarkDown rules for spacing
around the table has not been followed, making the table part of the
preceding text section, without the required space to be converted
properly to HTML on the docs.microsoft.com page.

Thanks to k10blogger (Siddharth Kaul) for reporting this issue.

Changes proposed:

- Add 2 blank lines before the table (at least 1 is required)
- Add a blank line before & after the page title (recommended)
- Add 1 blank space between each table cell divider (pipe character)
- Add a NewLine (line break) at EOF (end-of-file)

Ticket closure or reference:

Closes #4316