TLSExtensions xmlns Incorrectly contains "HTTPS"
While attempting to test SSO with always-on VPN I kept running into issues with the profile failing to be installed through both Intune and PowerShell. I compared a working manual configuration against the example code showed the only difference was the URI in the TLSExtensions. They should be "http://..." instead of "https://..."