IIS on Nano Server: HTTPS binding syntax
Description:

As reported in issue ticket #4320 (Incorrect syntax) and documented in
the MS Docs PowerShell scripting page "New-IISSite", some of the PS
cmdlet parameters need to be surrounded by double quotes in order to be
read correctly by the PowerShell command parser and applied to the
Nano Server IIS settings.

Thanks to @Famartincfm (Fausto Martins) for reporting this issue.

Changes proposed:
- Add double quotes around the following parameters:
    - Default Web Site
    - *:443:
    - My
    - 0
    - *:443:www.foo.bar.com
    - SNI
- Add "PowerShell" as MD syntax highlighting indicator to PS cmdlets
- Whitespace corrections:
    - Add newline (line break) at end of file
    - Remove redundant EOL (end-of-line) blanks
    - Add blank line between metadata and the page title
    - Add MarkDown indent marker compatibility spacing for "Applies to:"
    - Normalize numbered list spacing to 1 space before the text
    - Reduce max number of contiguous blank lines from 4 to 2

Ticket closure or reference:

Closes #4320