About Limit number of connections policy
When RDS client session is blocked by the Limit number of connections policy, the following message will appear : 
"The number of connections to this computer is limited and all connections are in use right now. Try connecting later or contact your system administrator."
Users do not receive "Remote Desktop Service is currently busy" message as described in this docs. So I think we should remove this section.