Windows Commands/forfiles: corrections & codestyle
Description:

As reported in issue ticket #4198 (Examples missing quotes), the
examples in this page need to have their documented double quotes for
the entire text following a /C parameter used by the FORFILES command.

Thanks to uecasm (Gavin Lambert) for reporting this issue, as well as
the Argentinian (Es-Es) users and devcrowley in the duplicate ticket.

Changes proposed:

- Add double quotes around the entire string following the /C parameters
- Update "path" to pathname, according to the command help text
- Convert the parameter letters from lowercase to UPPERCASE
- Use "searchmask" instead of "search mask" in one example context.
- Add spacing to the MarkDown tables, for source text readability

The cosmetic changes are done to reduce the chance of ambiguity,
especially when using /C "cmd /c [ ... ]" as well as readability.

The main goal is to get it closer to the command-line help text.

Ticket closure or reference:

Closes #4198