RDP file settings/RemoteApp: expandworkingdir
Description:

As pointed out in issue ticket #4417
(parameter remoteapplicationexpandworkingdir seems wrong), the setting
does not include the same :x:values parameter as the other settings.
The table documents which values to be used with the setting, so it
does look like the setting is only missing its parameter part.

Thanks to MarcMQC for noticing and reporting this issue.

Changes proposed:
- Add ':i:value' at the end of the setting name to indicate usage
- Add a blank line between the metadata section and the page title

Ticket closure or reference:

Closes #4417