Windows Commands/xcopy: Example 9 missing a verb
Description:

As reported in issue ticket #4383 (Grammar), the description text for
Example 9 is missing a verb to make complete sense. There is also a
minor grammatical error in the comment following the example.

Thanks to S-List for noticing and reporting the issue.

Changes proposed:
- Add the missing verb "copies" to complete the sentence
- Change "copy" to "copies" (3rd person singular) in the comment below
- Whitespace changes:
    - Normalize MD bullet point list spacing to 1 (6 lines)
    - Add 1 blank line between the metadata section and the page title

Additional notes:
- Please notify me if "will copy" is preferred instead of "copies".

Ticket closure or reference:

Closes #4383